requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
xgboost==2.0.3
shap==0.45.1
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
redis==5.0.4
python-multipart==0.0.9
python-dateutil==2.9.0
